Refactor filter factory
