+++ src/MagicOnion.Server/Filters/IMagicOnionFilterFactory.cs
@@ -2,9 +2,18 @@ using System;
 
 namespace MagicOnion.Server.Filters
 {
+    /// <summary>
+    /// An interface for filter which can create an instance of executable filter.
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
     public interface IMagicOnionFilterFactory<T> : IMagicOnionOrderedFilter, IMagicOnionFilterMetadata
         where T : IMagicOnionFilterMetadata
     {
+        /// <summary>
+        /// Creates an instance of executable filter.
+        /// </summary>
+        /// <param name="serviceLocator"></param>
+        /// <returns></returns>
         T CreateInstance(IServiceProvider serviceLocator);
     }
