+++ src/MagicOnion.Server/Filters/IMagicOnionFilterMetadata.cs
@@ -1,5 +1,8 @@
 namespace MagicOnion.Server.Filters;
 
+/// <summary>
+/// Marker interface for filters in MagicOnion process pipeline.
+/// </summary>
 public interface IMagicOnionFilterMetadata
 {
 }
