+++ src/MagicOnion.Server/Filters/IMagicOnionServiceFilter.cs
@@ -3,6 +3,9 @@ using System.Threading.Tasks;
 
 namespace MagicOnion.Server.Filters;
 
+/// <summary>
+/// An filter that surrounds execution of the Unary service method.
+/// </summary>
 public interface IMagicOnionServiceFilter : IMagicOnionFilterMetadata
 {
     ValueTask Invoke(ServiceContext context, Func<ServiceContext, ValueTask> next);
