+++ src/MagicOnion.Server/MethodHandler.cs
@@ -12,6 +12,7 @@ using System.Threading.Tasks;
 using MagicOnion.Internal;
 using Microsoft.Extensions.DependencyInjection;
 using MagicOnion.Server.Filters;
+using MagicOnion.Server.Filters.Internal;
 
 namespace MagicOnion.Server
 {
@@ -30,14 +31,11 @@ namespace MagicOnion.Server
 
         public ILookup<Type, Attribute> AttributeLookup { get; private set; }
 
-        readonly IMagicOnionFilterFactory<IMagicOnionServiceFilter>[] filters;
-
         // options
 
         internal readonly bool isReturnExceptionStackTraceInErrorDetail;
         internal readonly IMagicOnionLogger logger;
         readonly bool enableCurrentContext;
-        readonly IServiceProvider serviceProvider;
 
         // use for request handling.
 
@@ -57,7 +55,6 @@ namespace MagicOnion.Server
         public MethodHandler(Type classType, MethodInfo methodInfo, string methodName, MethodHandlerOptions handlerOptions, IServiceProvider serviceProvider)
         {
             this.methodHandlerId = Interlocked.Increment(ref methodHandlerIdBuild);
-            this.serviceProvider = serviceProvider;
 
             var serviceInterfaceType = classType.GetInterfaces().First(x => x.GetTypeInfo().IsGenericType && x.GetGenericTypeDefinition() == typeof(IService<>)).GetGenericArguments()[0];
 
@@ -85,12 +82,7 @@ namespace MagicOnion.Server
                 .Cast<Attribute>()
                 .ToLookup(x => x.GetType());
 
-            this.filters = handlerOptions.GlobalFilters
-                .OfType<IMagicOnionFilterFactory<IMagicOnionServiceFilter>>()
-                .Concat(classType.GetCustomAttributes(inherit: true).OfType<IMagicOnionFilterFactory<IMagicOnionServiceFilter>>().Select(x => new MagicOnionServiceFilterDescriptor(x, x.Order)))
-                .Concat(methodInfo.GetCustomAttributes(inherit: true).OfType<IMagicOnionFilterFactory<IMagicOnionServiceFilter>>().Select(x => new MagicOnionServiceFilterDescriptor(x, x.Order)))
-                .OrderBy(x => x.Order)
-                .ToArray();
+            var filterFactories = FilterHelper.GetFilterFactories(handlerOptions.GlobalFilters, classType, methodInfo);
 
             // options
             this.isReturnExceptionStackTraceInErrorDetail = handlerOptions.IsReturnExceptionStackTraceInErrorDetail;
@@ -143,7 +135,7 @@ namespace MagicOnion.Server
                         var body = Expression.Block(new[] { requestArg }, assignRequest, callBody);
                         var compiledBody = Expression.Lambda(body, contextArg).Compile();
 
-                        this.methodBody = BuildMethodBodyWithFilter((Func<ServiceContext, ValueTask>)compiledBody);
+                        this.methodBody = FilterHelper.WrapMethodBodyWithFilter(serviceProvider, filterFactories, (Func<ServiceContext, ValueTask>)compiledBody);
                     }
                     catch (Exception ex)
                     {
@@ -204,7 +196,7 @@ namespace MagicOnion.Server
                         var body = Expression.Block(new[] { requestArg }, assignRequest, callBody);
                         var compiledBody = Expression.Lambda(body, contextArg).Compile();
 
-                        this.methodBody = BuildMethodBodyWithFilter((Func<ServiceContext, ValueTask>)compiledBody);
+                        this.methodBody = FilterHelper.WrapMethodBodyWithFilter(serviceProvider, filterFactories, (Func<ServiceContext, ValueTask>)compiledBody);
                     }
                     catch (Exception ex)
                     {
@@ -249,7 +241,7 @@ namespace MagicOnion.Server
 
                         var compiledBody = Expression.Lambda(body, contextArg).Compile();
 
-                        this.methodBody = BuildMethodBodyWithFilter((Func<ServiceContext, ValueTask>)compiledBody);
+                        this.methodBody = FilterHelper.WrapMethodBodyWithFilter(serviceProvider, filterFactories, (Func<ServiceContext, ValueTask>)compiledBody);
                     }
                     catch (Exception ex)
                     {
@@ -320,19 +312,6 @@ namespace MagicOnion.Server
             {
                 throw new Exception($"Invalid return type, path:{methodInfo.DeclaringType!.Name + "/" + methodInfo.Name} type:{methodInfo.ReturnType.Name}");
             }
-        }
-
-        Func<ServiceContext, ValueTask> BuildMethodBodyWithFilter(Func<ServiceContext, ValueTask> methodBody)
-        {
-            Func<ServiceContext, ValueTask> next = methodBody;
-
-            foreach (var filterFactory in this.filters.Reverse())
-            {
-                var newFilter = filterFactory.CreateInstance(serviceProvider);
-                next = new InvokeHelper<ServiceContext, Func<ServiceContext, ValueTask>>(newFilter.Invoke, next).GetDelegate();
-            }
-
-            return next;
         }
 
         void BindUnaryHandler(ServiceBinderBase binder)
